+++ Apps/Contoso.WPF.Puppet.DotNetCore/MainWindow.xaml.cs
@@ -182,7 +182,7 @@ namespace Contoso.WPF.Puppet.DotNetCore
         private void SaveStorageSize_Click(object sender, EventArgs e)
         {
             var storageSize = StorageMaxSize.Text;
-            var size = (long)10 * 1024 * 1024;
+            var size = 10L * 1024 * 1024;
             long.TryParse(storageSize, out size);
             AppCenter.SetMaxStorageSizeAsync(size);
             Settings.Default.StorageMaxSize = size;
